pyproject.toml
[tool.poetry]
name = "qqbot"
version = "0.1.0"
description = ""
authors = ["euyo <xryazi1104@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = {extras = ["fastapi"], version = "^2.1.3"}
nonebot-adapter-onebot = "^2.3.1"
nonebot-plugin-apscheduler = "^0.3.0"
httpx = "^0.26.0"
requests = "^2.31.0"
pinyin = "^0.4.0"
jieba = "^0.42.1"
nonebot-plugin-htmlrender = "^0.2.3"
aiocache = "^0.12.2"
peewee = "^3.17.0"
pillow = "^9.2.0"
nonebot-plugin-localstore = "^0.5.1"
nonebot-plugin-session-saa = "^0.1.1"
emoji = "^2.9.0"
aiohttp = "^3.9.1"
lxml = "^5.1.0"
setuptools = "^69.0.3"
tqdm = "^4.66.1"
meme-generator = "^0.0.17"
pypinyin = "^0.50.0"
nonebot-plugin-userinfo = "^0.1.3"
nonebot-plugin-alconna = "^0.35.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "ali"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"


[[tool.poetry.source]]
name = "tencent"
url = "https://mirrors.cloud.tencent.com/pypi/simple/"
priority = "primary"


[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
priority = "default"
